Rename 'engine' library to less ambigous 'graphics'.
