+++ src/CMakeLists.txt
@@ -5,7 +5,7 @@ add_subdirectory(lib/GSL-2.0.0/gsl)
 add_subdirectory(lib/arte/arte)
 add_subdirectory(lib/renderer/renderer)
 add_subdirectory(lib/resource/resource)
-add_subdirectory(lib/engine/engine)
+add_subdirectory(lib/graphics/graphics)
 
 option(BUILD_tests "Build the demo applications" ON)
 if(BUILD_tests)
