+++ src/app/prototypes/bloom/CMakeLists.txt
@@ -20,7 +20,7 @@ add_executable(${TARGET_NAME}
 target_link_libraries(${TARGET_NAME}
     PRIVATE
         glad::glad
-        ad::engine
+        ad::graphics
         ad::renderer
         ad::resource
 )
