+++ src/app/prototypes/chader/CMakeLists.txt
@@ -25,7 +25,7 @@ add_executable(${TARGET_NAME}
 target_link_libraries(${TARGET_NAME}
     PRIVATE
         glad::glad
-        ad::engine
+        ad::graphics
         ad::renderer
         gsl::gsl
         Boost::filesystem
