+++ src/app/prototypes/curve/Scene.h
@@ -3,7 +3,7 @@
 
 #include "shaders.h"
 
-#include <engine/Timer.h>
+#include <graphics/Timer.h>
 
 #include <renderer/Drawing.h>
 #include <renderer/Uniforms.h>
