+++ src/app/prototypes/polyline/CMakeLists.txt
@@ -24,7 +24,7 @@ target_link_libraries(${TARGET_NAME}
         glad::glad
         glfw
 
-        ad::engine
+        ad::graphics
         ad::math
         ad::renderer
 )
