+++ src/app/prototypes/tiler/CMakeLists.txt
@@ -18,9 +18,9 @@ add_executable(${TARGET_NAME}
 
 target_link_libraries(${TARGET_NAME}
     PRIVATE
-        glfw
-        engine
-        resource
+        glfw::glfw
+        ad::graphics
+        ad::resource
         Boost::boost # Target for header-only dependencies
 )
