+++ src/lib/graphics/graphics/Curving.cpp
