+++ src/lib/graphics/graphics/Curving.h
