+++ src/lib/graphics/graphics/TrivialLineStrip.cpp
