+++ src/lib/graphics/graphics/TrivialLineStrip.h
