+++ src/lib/graphics/graphics/TrivialShaping.cpp
