+++ src/lib/graphics/graphics/commons.h
